+++ ChangeLog.md
@@ -5,6 +5,7 @@ XP AST ChangeLog
 
 ## 7.0.0 / ????-??-??
 
+* Merged PR #20: Remove transformations API - @thekid
 * Merged PR #17: Refactor parsing to allow blocks anywhere an expression
   is allowed. This not only allows `fn() => { ... }` but also using blocks
   in `match` expressions.